chore: strip C++ bridges from Rust crates, remove kofem-wasm/kofem-py from workspace
The C++ engine (engine/) is now the only WASM build path; the Rust-to-C++
FFI bridges in kofem-geom/kofem-mesh/kofem-core were duplicating that work
and forcing native OCCT/Netgen/MFEM installs in CI.

Removes:
- build.rs, cpp/, include/ from kofem-geom, kofem-mesh, kofem-core
- kofem-core/tests/pipeline.rs (tested the C++ bridges)
- kofem-wasm and kofem-py from workspace members (still in tree, not built)

Keeps pure-Rust types and traits (SurfaceMesh, VolumeMesh, MeshOptions,
BoundaryConditions, LinearElasticMaterial, FemSolver) as the shared data
layer for future Rust implementations.
